Changed how the potions work

+++ RGProject/Items/Potion.cs
@@ -7,6 +7,6 @@ public class Potion() : Item
 	public Guid Id { get; } = Guid.NewGuid();
 	public string Name { get; set; }
 	public string Description { get; set; }
-	public int Power { get; set; }
+	public double Power { get; set; }
 	public PotionModifier Stat { get; set; }
 }
